Convert testData to expected line endings
git will convert line endings for given file types accordingly to rules
defined in gitattribues file. This was not desired for a test data
embedded in the unit test script, which defined a piece of text data
with dos line endings. After commit the line endings have been converted
to unix ones, in result causing the test to fail. This change fixes the
issue by post processing the data to expected format.